- Rename `structs` folder to `struct` - Include files in `cglms.h` - Fix style issues (tabs to spaces, open brace without newline)
