+++ include/cglm/struct/frustum.h
@@ -55,9 +55,9 @@
 CGLM_INLINE
 void
 glms_frustum_planes(mat4s m, vec4s dest[6]) {
-	vec4 rawDest[6];
-	glm_frustum_planes(m.raw, rawDest);
-	glms_vec4_pack(dest, rawDest, 6);
+  vec4 rawDest[6];
+  glm_frustum_planes(m.raw, rawDest);
+  glms_vec4_pack(dest, rawDest, 6);
 }
 
 /*!
@@ -86,9 +86,9 @@ glms_frustum_planes(mat4s m, vec4s dest[6]) {
 CGLM_INLINE
 void
 glms_frustum_corners(mat4s invMat, vec4s dest[8]) {
-	vec4 rawDest[8];
-	glm_frustum_corners(invMat.raw, rawDest);
-	glms_vec4_pack(dest, rawDest, 8);
+  vec4 rawDest[8];
+  glm_frustum_corners(invMat.raw, rawDest);
+  glms_vec4_pack(dest, rawDest, 8);
 }
 
 /*!
@@ -100,12 +100,12 @@ glms_frustum_corners(mat4s invMat, vec4s dest[8]) {
 CGLM_INLINE
 vec4s
 glms_frustum_center(vec4s corners[8]) {
-	vec4 rawCorners[8];
-	vec4s r;
+  vec4 rawCorners[8];
+  vec4s r;
 
-	glms_vec4_unpack(rawCorners, corners, 8);
-	glm_frustum_center(rawCorners, r.raw);
-	return r;
+  glms_vec4_unpack(rawCorners, corners, 8);
+  glm_frustum_center(rawCorners, r.raw);
+  return r;
 }
 
 /*!
@@ -118,12 +118,12 @@ glms_frustum_center(vec4s corners[8]) {
 CGLM_INLINE
 void
 glms_frustum_box(vec4s corners[8], mat4s m, vec3s box[2]) {
-	vec4 rawCorners[8];
-	vec3 rawBox[2];
+  vec4 rawCorners[8];
+  vec3 rawBox[2];
 
-	glms_vec4_unpack(rawCorners, corners, 8);
-	glm_frustum_box(rawCorners, m.raw, rawBox);
-	glms_vec3_pack(box, rawBox, 2);
+  glms_vec4_unpack(rawCorners, corners, 8);
+  glm_frustum_box(rawCorners, m.raw, rawBox);
+  glms_vec3_pack(box, rawBox, 2);
 }
 
 /*!
@@ -144,12 +144,12 @@ glms_frustum_corners_at(vec4s corners[8],
                         float splitDist,
                         float farDist,
                         vec4s planeCorners[4]) {
-	vec4 rawCorners[8];
-	vec4 rawPlaneCorners[4];
+  vec4 rawCorners[8];
+  vec4 rawPlaneCorners[4];
 
-	glms_vec4_unpack(rawCorners, corners, 8);
-	glm_frustum_corners_at(rawCorners, splitDist, farDist, rawPlaneCorners);
-	glms_vec4_pack(planeCorners, rawPlaneCorners, 8);
+  glms_vec4_unpack(rawCorners, corners, 8);
+  glm_frustum_corners_at(rawCorners, splitDist, farDist, rawPlaneCorners);
+  glms_vec4_pack(planeCorners, rawPlaneCorners, 8);
 }
 
 #endif /* cglm_frustums_h */
