+++ include/cglm/struct/mat4.h
@@ -241,11 +241,11 @@ CGLM_INLINE
 mat4s
 glms_mat4_mulN(mat4s * __restrict matrices[], uint32_t len) {
   CGLM_ALIGN_MAT mat4s r = GLMS_MAT4_IDENTITY_INIT;
-	size_t i;
+  size_t i;
 
   for (i = 0; i < len; i++) {
-		r = glms_mat4_mul(r, *matrices[i]);
-	}
+    r = glms_mat4_mul(r, *matrices[i]);
+  }
   
   return r;
 }
