+++ include/cglm/struct/plane.h
@@ -33,8 +33,8 @@
 CGLM_INLINE
 vec4s
 glms_plane_normalize(vec4s plane) {
-	glm_plane_normalize(plane.raw);
-	return plane;
+  glm_plane_normalize(plane.raw);
+  return plane;
 }
 
 #endif /* cglm_planes_h */
