+++ include/cglm/struct/vec3.h
@@ -17,53 +17,53 @@
 
  Functions:
    CGLM_INLINE vec3s glms_vec3(vec4s v4);
-	 CGLM_INLINE vec3s glms_vec3_copy(vec3s v);
-	 CGLM_INLINE void  glms_vec3_pack(vec3s dst[], vec3 src[], size_t len);
-	 CGLM_INLINE void  glms_vec3_unpack(vec3 dst[], vec3s src[], size_t len);
-	 CGLM_INLINE vec3s glms_vec3_zero();
-	 CGLM_INLINE vec3s glms_vec3_one();
-	 CGLM_INLINE float glms_vec3_dot(vec3s a, vec3s b);
-	 CGLM_INLINE float glms_vec3_norm2(vec3s v);
-	 CGLM_INLINE float glms_vec3_norm(vec3s v);
-	 CGLM_INLINE vec3s glms_vec3_add(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_vec3_adds(vec3s a, float s);
-	 CGLM_INLINE vec3s glms_vec3_sub(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_vec3_subs(vec3s a, float s);
-	 CGLM_INLINE vec3s glms_vec3_mul(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_vec3_scale(vec3s v, float s);
-	 CGLM_INLINE vec3s glms_vec3_scale_as(vec3s v, float s);
-	 CGLM_INLINE vec3s glms_vec3_div(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_vec3_divs(vec3s a, float s);
-	 CGLM_INLINE vec3s glms_vec3_addadd(vec3s a, vec3s b, vec3s dest);
-	 CGLM_INLINE vec3s glms_vec3_subadd(vec3s a, vec3s b, vec3s dest);
-	 CGLM_INLINE vec3s glms_vec3_muladd(vec3s a, vec3s b, vec3s dest);
-	 CGLM_INLINE vec3s glms_vec3_muladds(vec3s a, float s, vec3s dest);
-	 CGLM_INLINE vec3s glms_vec3_maxadd(vec3s a, vec3s b, vec3s dest);
-	 CGLM_INLINE vec3s glms_vec3_minadd(vec3s a, vec3s b, vec3s dest);
-	 CGLM_INLINE vec3s glms_vec3_flipsign(vec3s v);
-	 CGLM_INLINE vec3s glms_vec3_negate(vec3s v);
-	 CGLM_INLINE vec3s glms_vec3_inv(vec3s v);
-	 CGLM_INLINE vec3s glms_vec3_normalize(vec3s v);
-	 CGLM_INLINE vec3s glms_vec3_cross(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_vec3_crossn(vec3s a, vec3s b);
-	 CGLM_INLINE float glms_vec3_distance(vec3s a, vec3s b);
-	 CGLM_INLINE float glms_vec3_angle(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_vec3_rotate(vec3s v, float angle, vec3s axis);
-	 CGLM_INLINE vec3s glms_vec3_rotate_m4(mat4s m, vec3s v);
-	 CGLM_INLINE vec3s glms_vec3_rotate_m3(mat3s m, vec3s v);
-	 CGLM_INLINE vec3s glms_vec3_proj(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_vec3_center(vec3s a, vec3s b);
-	 CGLM_INLINE float glms_vec3_distance2(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_vec3_maxv(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_vec3_minv(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_vec3_ortho(vec3s v);
-	 CGLM_INLINE vec3s glms_vec3_clamp(vec3s v, float minVal, float maxVal);
-	 CGLM_INLINE vec3s glms_vec3_lerp(vec3s from, vec3s to, float t);
+   CGLM_INLINE vec3s glms_vec3_copy(vec3s v);
+   CGLM_INLINE void  glms_vec3_pack(vec3s dst[], vec3 src[], size_t len);
+   CGLM_INLINE void  glms_vec3_unpack(vec3 dst[], vec3s src[], size_t len);
+   CGLM_INLINE vec3s glms_vec3_zero();
+   CGLM_INLINE vec3s glms_vec3_one();
+   CGLM_INLINE float glms_vec3_dot(vec3s a, vec3s b);
+   CGLM_INLINE float glms_vec3_norm2(vec3s v);
+   CGLM_INLINE float glms_vec3_norm(vec3s v);
+   CGLM_INLINE vec3s glms_vec3_add(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_vec3_adds(vec3s a, float s);
+   CGLM_INLINE vec3s glms_vec3_sub(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_vec3_subs(vec3s a, float s);
+   CGLM_INLINE vec3s glms_vec3_mul(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_vec3_scale(vec3s v, float s);
+   CGLM_INLINE vec3s glms_vec3_scale_as(vec3s v, float s);
+   CGLM_INLINE vec3s glms_vec3_div(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_vec3_divs(vec3s a, float s);
+   CGLM_INLINE vec3s glms_vec3_addadd(vec3s a, vec3s b, vec3s dest);
+   CGLM_INLINE vec3s glms_vec3_subadd(vec3s a, vec3s b, vec3s dest);
+   CGLM_INLINE vec3s glms_vec3_muladd(vec3s a, vec3s b, vec3s dest);
+   CGLM_INLINE vec3s glms_vec3_muladds(vec3s a, float s, vec3s dest);
+   CGLM_INLINE vec3s glms_vec3_maxadd(vec3s a, vec3s b, vec3s dest);
+   CGLM_INLINE vec3s glms_vec3_minadd(vec3s a, vec3s b, vec3s dest);
+   CGLM_INLINE vec3s glms_vec3_flipsign(vec3s v);
+   CGLM_INLINE vec3s glms_vec3_negate(vec3s v);
+   CGLM_INLINE vec3s glms_vec3_inv(vec3s v);
+   CGLM_INLINE vec3s glms_vec3_normalize(vec3s v);
+   CGLM_INLINE vec3s glms_vec3_cross(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_vec3_crossn(vec3s a, vec3s b);
+   CGLM_INLINE float glms_vec3_distance(vec3s a, vec3s b);
+   CGLM_INLINE float glms_vec3_angle(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_vec3_rotate(vec3s v, float angle, vec3s axis);
+   CGLM_INLINE vec3s glms_vec3_rotate_m4(mat4s m, vec3s v);
+   CGLM_INLINE vec3s glms_vec3_rotate_m3(mat3s m, vec3s v);
+   CGLM_INLINE vec3s glms_vec3_proj(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_vec3_center(vec3s a, vec3s b);
+   CGLM_INLINE float glms_vec3_distance2(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_vec3_maxv(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_vec3_minv(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_vec3_ortho(vec3s v);
+   CGLM_INLINE vec3s glms_vec3_clamp(vec3s v, float minVal, float maxVal);
+   CGLM_INLINE vec3s glms_vec3_lerp(vec3s from, vec3s to, float t);
 
  Convenient:
    CGLM_INLINE vec3s glms_cross(vec3s a, vec3s b);
-	 CGLM_INLINE float glms_dot(vec3s a, vec3s b);
-	 CGLM_INLINE vec3s glms_normalize(vec3s v);
+   CGLM_INLINE float glms_dot(vec3s a, vec3s b);
+   CGLM_INLINE vec3s glms_normalize(vec3s v);
  */
 
 #ifndef cglm_vec3s_h
@@ -89,7 +89,7 @@
  * @brief init vec3 using vec4
  *
  * @param[in]  v4   vector4
- * @returns				  destination
+ * @returns         destination
  */
 CGLM_INLINE
 vec3s
@@ -103,7 +103,7 @@ glms_vec3(vec4s v4) {
  * @brief copy all members of [a] to [dest]
  *
  * @param[in]  a    source
- * @returns					destination
+ * @returns         destination
  */
 CGLM_INLINE
 vec3s
@@ -123,13 +123,13 @@ glms_vec3_copy(vec3s v) {
 CGLM_INLINE
 void
 glms_vec3_pack(vec3s dst[], vec3 src[], size_t len) {
-	size_t i;
-	
-	for (i = 0; i < len; i++) {
-		dst[i].x = src[i][0];
-		dst[i].y = src[i][1];
-		dst[i].z = src[i][2];
-	}
+  size_t i;
+
+  for (i = 0; i < len; i++) {
+    dst[i].x = src[i][0];
+    dst[i].y = src[i][1];
+    dst[i].z = src[i][2];
+  }
 }
 
 /*!
@@ -142,20 +142,20 @@ glms_vec3_pack(vec3s dst[], vec3 src[], size_t len) {
 CGLM_INLINE
 void
 glms_vec3_unpack(vec3 dst[], vec3s src[], size_t len) {
-	size_t i;
+  size_t i;
 
-	for (i = 0; i < len; i++) {
-		dst[i][0] = src[i].x;
-		dst[i][1] = src[i].y;
-		dst[i][2] = src[i].z;
-	}	
+  for (i = 0; i < len; i++) {
+    dst[i][0] = src[i].x;
+    dst[i][1] = src[i].y;
+    dst[i][2] = src[i].z;
+  }
 }
 
 /*!
  * @brief make vector zero
  *
- * @param[in]  v vector
- * @returns			 zero vector
+ * @param[in]  v  vector
+ * @returns       zero vector
  */
 CGLM_INLINE
 vec3s
@@ -168,8 +168,8 @@ glms_vec3_zero() {
 /*!
  * @brief make vector one
  *
- * @param[in]  v vector
- * @returns 		 one vector
+ * @param[in]  v  vector
+ * @returns       one vector
  */
 CGLM_INLINE
 vec3s
@@ -228,7 +228,7 @@ glms_vec3_norm(vec3s v) {
  *
  * @param[in]  a    vector1
  * @param[in]  b    vector2
- * @returns				  destination vector
+ * @returns         destination vector
  */
 CGLM_INLINE
 vec3s
@@ -243,7 +243,7 @@ glms_vec3_add(vec3s a, vec3s b) {
  *
  * @param[in]  v    vector
  * @param[in]  s    scalar
- * @returns					destination vector
+ * @returns         destination vector
  */
 CGLM_INLINE
 vec3s
@@ -258,7 +258,7 @@ glms_vec3_adds(vec3s a, float s) {
  *
  * @param[in]  a    vector1
  * @param[in]  b    vector2
- * @returns 				destination vector
+ * @returns         destination vector
  */
 CGLM_INLINE
 vec3s
@@ -273,7 +273,7 @@ glms_vec3_sub(vec3s a, vec3s b) {
  *
  * @param[in]  v    vector
  * @param[in]  s    scalar
- * @returns 				destination vector
+ * @returns         destination vector
  */
 CGLM_INLINE
 vec3s
@@ -286,9 +286,9 @@ glms_vec3_subs(vec3s a, float s) {
 /*!
  * @brief multiply two vector (component-wise multiplication)
  *
- * @param a    vector1
- * @param b    vector2
- * @returns		 v3 = (a[0] * b[0], a[1] * b[1], a[2] * b[2])
+ * @param     a     vector1
+ * @param     b     vector2
+ * @returns         v3 = (a[0] * b[0], a[1] * b[1], a[2] * b[2])
  */
 CGLM_INLINE
 vec3s
@@ -303,7 +303,7 @@ glms_vec3_mul(vec3s a, vec3s b) {
  *
  * @param[in]  v    vector
  * @param[in]  s    scalar
- * @returns 				destination vector
+ * @returns         destination vector
  */
 CGLM_INLINE
 vec3s
@@ -318,7 +318,7 @@ glms_vec3_scale(vec3s v, float s) {
  *
  * @param[in]  v    vector
  * @param[in]  s    scalar
- * @returns 				destination vector
+ * @returns         destination vector
  */
 CGLM_INLINE
 vec3s
@@ -333,7 +333,7 @@ glms_vec3_scale_as(vec3s v, float s) {
  *
  * @param[in]  a    vector 1
  * @param[in]  b    vector 2
- * @returns					result = (a[0]/b[0], a[1]/b[1], a[2]/b[2])
+ * @returns         result = (a[0]/b[0], a[1]/b[1], a[2]/b[2])
  */
 CGLM_INLINE
 vec3s
@@ -348,7 +348,7 @@ glms_vec3_div(vec3s a, vec3s b) {
  *
  * @param[in]  v    vector
  * @param[in]  s    scalar
- * @returns					result = (a[0]/s, a[1]/s, a[2]/s)
+ * @returns         result = (a[0]/s, a[1]/s, a[2]/s)
  */
 CGLM_INLINE
 vec3s
@@ -365,7 +365,7 @@ glms_vec3_divs(vec3s a, float s) {
  *
  * @param[in]  a    vector 1
  * @param[in]  b    vector 2
- * @returns 				dest += (a + b)
+ * @returns         dest += (a + b)
  */
 CGLM_INLINE
 vec3s
@@ -381,7 +381,7 @@ glms_vec3_addadd(vec3s a, vec3s b, vec3s dest) {
  *
  * @param[in]  a    vector 1
  * @param[in]  b    vector 2
- * @returns					dest += (a + b)
+ * @returns         dest += (a + b)
  */
 CGLM_INLINE
 vec3s
@@ -397,7 +397,7 @@ glms_vec3_subadd(vec3s a, vec3s b, vec3s dest) {
  *
  * @param[in]  a    vector 1
  * @param[in]  b    vector 2
- * @returns					dest += (a * b)
+ * @returns         dest += (a * b)
  */
 CGLM_INLINE
 vec3s
@@ -413,7 +413,7 @@ glms_vec3_muladd(vec3s a, vec3s b, vec3s dest) {
  *
  * @param[in]  a    vector
  * @param[in]  s    scalar
- * @returns 				dest += (a * b)
+ * @returns         dest += (a * b)
  */
 CGLM_INLINE
 vec3s
@@ -429,7 +429,7 @@ glms_vec3_muladds(vec3s a, float s, vec3s dest) {
  *
  * @param[in]  a    vector 1
  * @param[in]  b    vector 2
- * @returns					dest += max(a, b)
+ * @returns         dest += max(a, b)
  */
 CGLM_INLINE
 vec3s
@@ -445,7 +445,7 @@ glms_vec3_maxadd(vec3s a, vec3s b, vec3s dest) {
  *
  * @param[in]  a    vector 1
  * @param[in]  b    vector 2
- * @returns					dest += min(a, b)
+ * @returns         dest += min(a, b)
  */
 CGLM_INLINE
 vec3s
@@ -458,7 +458,7 @@ glms_vec3_minadd(vec3s a, vec3s b, vec3s dest) {
  * @brief negate vector components and store result in dest
  *
  * @param[in]   v     vector
- * @returns						result vector
+ * @returns           result vector
  */
 CGLM_INLINE
 vec3s
@@ -471,7 +471,7 @@ glms_vec3_flipsign(vec3s v) {
  * @brief negate vector components
  *
  * @param[in]  v  vector
- * @returns 			negated vector
+ * @returns       negated vector
  */
 CGLM_INLINE
 vec3s
@@ -484,7 +484,7 @@ glms_vec3_negate(vec3s v) {
  * @brief normalize vec3 and store result in same vec
  *
  * @param[in] v vector
- * @returns			normalized vector
+ * @returns     normalized vector
  */
 CGLM_INLINE
 vec3s
@@ -498,7 +498,7 @@ glms_vec3_normalize(vec3s v) {
  *
  * @param[in]  a    vector 1
  * @param[in]  b    vector 2
- * @returns					destination
+ * @returns         destination
  */
 CGLM_INLINE
 vec3s
@@ -513,7 +513,7 @@ glms_vec3_cross(vec3s a, vec3s b) {
  *
  * @param[in]  a    vector 1
  * @param[in]  b    vector 2
- * @returns					destination
+ * @returns         destination
  */
 CGLM_INLINE
 vec3s
@@ -540,10 +540,10 @@ glms_vec3_angle(vec3s a, vec3s b) {
 /*!
  * @brief rotate vec3 around axis by angle using Rodrigues' rotation formula
  *
- * @param[in] 		 v     vector
- * @param[in]      axis  axis vector (must be unit vector)
- * @param[in]      angle angle by radians
- * @returns							 rotated vector
+ * @param[in]     v     vector
+ * @param[in]     axis  axis vector (must be unit vector)
+ * @param[in]     angle angle by radians
+ * @returns             rotated vector
  */
 CGLM_INLINE
 vec3s
@@ -563,7 +563,7 @@ glms_vec3_rotate(vec3s v, float angle, vec3s axis) {
  *
  * @param[in]  m    affine matrix or rot matrix
  * @param[in]  v    vector
- * @returns					rotated vector
+ * @returns         rotated vector
  */
 CGLM_INLINE
 vec3s
@@ -578,7 +578,7 @@ glms_vec3_rotate_m4(mat4s m, vec3s v) {
  *
  * @param[in]  m    affine matrix or rot matrix
  * @param[in]  v    vector
- * @returns					rotated vector
+ * @returns         rotated vector
  */
 CGLM_INLINE
 vec3s
@@ -593,7 +593,7 @@ glms_vec3_rotate_m3(mat3s m, vec3s v) {
  *
  * @param[in]  a    vector1
  * @param[in]  b    vector2
- * @returns 				projected vector
+ * @returns         projected vector
  */
 CGLM_INLINE
 vec3s
@@ -608,7 +608,7 @@ glms_vec3_proj(vec3s a, vec3s b) {
  *
  * @param[in]  a    vector1
  * @param[in]  b    vector2
- * @returns				  center point
+ * @returns         center point
  */
 CGLM_INLINE
 vec3s
@@ -623,7 +623,7 @@ glms_vec3_center(vec3s a, vec3s b) {
  *
  * @param[in] a vector1
  * @param[in] b vector2
- * @return 			squared distance (distance * distance)
+ * @return      squared distance (distance * distance)
  */
 CGLM_INLINE
 float
@@ -636,7 +636,7 @@ glms_vec3_distance2(vec3s a, vec3s b) {
  *
  * @param[in] a vector1
  * @param[in] b vector2
- * @return 			distance
+ * @return      distance
  */
 CGLM_INLINE
 float
@@ -649,7 +649,7 @@ glms_vec3_distance(vec3s a, vec3s b) {
  *
  * @param[in]  a    vector1
  * @param[in]  b    vector2
- * @returns					destination
+ * @returns         destination
  */
 CGLM_INLINE
 vec3s
@@ -664,7 +664,7 @@ glms_vec3_maxv(vec3s a, vec3s b) {
  *
  * @param[in]  a    vector1
  * @param[in]  b    vector2
- * @returns					destination
+ * @returns         destination
  */
 CGLM_INLINE
 vec3s
@@ -678,7 +678,7 @@ glms_vec3_minv(vec3s a, vec3s b) {
  * @brief possible orthogonal/perpendicular vector
  *
  * @param[in]  v    vector
- * @returns					orthogonal/perpendicular vector
+ * @returns         orthogonal/perpendicular vector
  */
 CGLM_INLINE
 vec3s
@@ -691,10 +691,10 @@ glms_vec3_ortho(vec3s v) {
 /*!
  * @brief clamp vector's individual members between min and max values
  *
- * @param[in]  		  v      vector
- * @param[in]       minVal minimum value
- * @param[in]       maxVal maximum value
- * @returns								 clamped vector
+ * @param[in]       v       vector
+ * @param[in]       minVal  minimum value
+ * @param[in]       maxVal  maximum value
+ * @returns                 clamped vector
  */
 CGLM_INLINE
 vec3s
@@ -708,10 +708,10 @@ glms_vec3_clamp(vec3s v, float minVal, float maxVal) {
  *
  * formula:  from + s * (to - from)
  *
- * @param[in]   from from value
- * @param[in]   to   to value
- * @param[in]   t    interpolant (amount) clamped between 0 and 1
- * @returns					 destination
+ * @param[in]   from  from value
+ * @param[in]   to    to value
+ * @param[in]   t     interpolant (amount) clamped between 0 and 1
+ * @returns           destination
  */
 CGLM_INLINE
 vec3s
@@ -726,9 +726,9 @@ glms_vec3_lerp(vec3s from, vec3s to, float t) {
  *
  * this is just convenient wrapper
  *
- * @param[in]  a source 1
- * @param[in]  b source 2
- * @returns			 destination
+ * @param[in]  a  source 1
+ * @param[in]  b  source 2
+ * @returns       destination
  */
 CGLM_INLINE
 vec3s
@@ -745,8 +745,7 @@ glms_cross(vec3s a, vec3s b) {
  *
  * @param[in] a vector1
  * @param[in] b vector2
- *
- * @return 		  dot product
+ * @return      dot product
  */
 CGLM_INLINE
 float
@@ -759,8 +758,8 @@ glms_dot(vec3s a, vec3s b) {
  *
  * this is just convenient wrapper
  *
- * @param[in] 	v 	 vector
- * @returns					 normalized vector
+ * @param[in]   v   vector
+ * @returns         normalized vector
  */
 CGLM_INLINE
 vec3s
