+++ test/src/test_common.c
@@ -49,8 +49,7 @@ test_rand_vec3(vec3 dest) {
 }
 
 vec3s
-test_rand_vec3s()
-{
+test_rand_vec3s() {
   vec3s r;
   test_rand_vec3(r.raw);
   return r;
@@ -67,8 +66,7 @@ test_rand_vec4(vec4 dest) {
 }
 
 vec4s
-test_rand_vec4s()
-{
+test_rand_vec4s() {
   vec4s r;
   test_rand_vec4(r.raw);
   return r;
@@ -137,8 +135,7 @@ test_assert_vec3_eq(vec3 v1, vec3 v2) {
 }
 
 void
-test_assert_vec3s_eq(vec3s v1, vec3s v2)
-{
+test_assert_vec3s_eq(vec3s v1, vec3s v2) {
   test_assert_vec3_eq(v1.raw, v2.raw);
 }
 
@@ -151,8 +148,7 @@ test_assert_vec4_eq(vec4 v1, vec4 v2) {
 }
 
 void
-test_assert_vec4s_eq(vec4s v1, vec4s v2)
-{
+test_assert_vec4s_eq(vec4s v1, vec4s v2) {
   test_assert_vec4_eq(v1.raw, v2.raw);
 }
